% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{addQuery}
\alias{addQuery}
\title{adds a new query to the plot}
\usage{
addQuery(upsetjs, name, color, elems = NULL, set = NULL)
}
\arguments{
\item{upsetjs}{the upsetjs (proxy) instance}

\item{name}{name of the query}

\item{color}{color of the query}

\item{elems}{the list of elems to highlight}

\item{set}{the set name, similar to the selection}
}
\description{
adds a new query to the plot
}
