% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{chartLayout}
\alias{chartLayout}
\title{specify the chart layout}
\usage{
chartLayout(
  upsetjs,
  height.ratios = NULL,
  width.ratios = NULL,
  padding = NULL,
  bar.padding = NULL
)
}
\arguments{
\item{upsetjs}{the upsetjs (proxy) instance}
}
\description{
specify the chart layout
}
