% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{chartTheme}
\alias{chartTheme}
\title{specify theming options}
\usage{
chartTheme(
  upsetjs,
  theme = NULL,
  selection.color = NULL,
  alternating.color = NULL,
  color = NULL,
  text.color = NULL,
  hover.hint.color = NULL,
  not.member.color = NULL
)
}
\arguments{
\item{upsetjs}{the upsetjs (proxy) instance}

\item{theme}{theme to use 'dark' or 'light'}

\item{selection.color}{selection color}

\item{alternating.color}{alternating background color}

\item{color}{main bar color}

\item{text.color}{main text color}

\item{hover.hint.color}{color of the hover hint}

\item{not.member.color}{color of the dot if not a member}
}
\value{
upsetjs
}
\description{
specify theming options
}
