% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{clearQueries}
\alias{clearQueries}
\title{clears the list of queries for incremental updates}
\usage{
clearQueries(upsetjs)
}
\arguments{
\item{upsetjs}{the upsetjs (proxy) instance}
}
\value{
upsetjs
}
\description{
clears the list of queries for incremental updates
}
