% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{fromDataFrame}
\alias{fromDataFrame}
\title{extract the sets from a data frame (rows = elems, columns = sets, cell = contained)}
\usage{
fromDataFrame(upsetjs, df, order.by = "freq")
}
\description{
extract the sets from a data frame (rows = elems, columns = sets, cell = contained)
}
