% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{fromExpression}
\alias{fromExpression}
\title{generates the sets from a lists object that contained the cardinalties of both sets and combinations (&)}
\usage{
fromExpression(upsetjs, value, symbol = "&", order.by = "cardinality")
}
\arguments{
\item{upsetjs}{the upsetjs (proxy) instance}

\item{value}{the expression list input}

\item{symbol}{the symbol how to split list names to get the sets}

\item{order.by}{order intersections by cardinality or degree}
}
\value{
upsetjs
}
\description{
generates the sets from a lists object that contained the cardinalties of both sets and combinations (&)
}
