% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{generateIntersections}
\alias{generateIntersections}
\title{generate intersections}
\usage{
generateIntersections(
  upsetjs,
  min = NULL,
  max = NULL,
  empty = NULL,
  order.by = "freq"
)
}
\arguments{
\item{upsetjs}{the upsetjs (proxy) instance}
}
\description{
generate intersections
}
