% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{heightRatios}
\alias{heightRatios}
\title{specify the ratios between the intersection and set plot}
\usage{
heightRatios(upsetjs, ratios)
}
\description{
specify the ratios between the intersection and set plot
}
