% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{queryLegend}
\alias{queryLegend}
\title{renders a legend for the queries}
\usage{
queryLegend(upsets, value = TRUE)
}
\arguments{
\item{value}{whether to enable or disable}

\item{upsetjs}{the upsetjs (proxy) instance}
}
\description{
renders a legend for the queries
}
