% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{renderUpsetjs}
\alias{renderUpsetjs}
\title{Shiny render bindings for upsetjs}
\usage{
renderUpsetjs(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that generates an upset}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Shiny render bindings for upsetjs
}
