% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{setQueries}
\alias{setQueries}
\title{set the queries}
\usage{
setQueries(upsetjs, queries = list())
}
\arguments{
\item{upsetjs}{the upsetjs (proxy) instance}

\item{queries}{the queries to set}
}
\value{
upsetjs
}
\description{
set the queries
}
