% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{setSelection}
\alias{setSelection}
\title{sets the selection of the chart}
\usage{
setSelection(upsetjs, name = NULL)
}
\arguments{
\item{upsetjs}{the upsetjs (proxy) instance}

\item{name}{the name of the set to select}
}
\value{
upsetjs
}
\description{
sets the selection of the chart
}
