% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{upsetjs-shiny}
\alias{upsetjs-shiny}
\alias{upsetjsOutput}
\alias{renderUpsetjs}
\title{Shiny bindings for upsetjs}
\usage{
upsetjsOutput(outputId, width = "100\%", height = "400px")

renderUpsetjs(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width}{Must be a valid CSS unit (like \code{'100\%'},
\code{'800px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{height}{see width}

\item{expr}{An expression that generates an upset}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using UpSet.js within Shiny
applications and interactive Rmd documents.
}
