% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsetjs.R
\name{upsetjs}
\alias{upsetjs}
\title{upsetjs - factory for UpSet HTMLWidget}
\usage{
upsetjs(
  width = "100\%",
  height = NULL,
  elementId = NULL,
  sizingPolicy = upsetjsSizingPolicy()
)
}
\arguments{
\item{width}{width of the element}

\item{height}{height of the element}

\item{elementId}{unique element id}

\item{sizingPolicy}{htmlwidgets sizing policy object. Defaults to \code{\link{upsetjsSizingPolicy}()}}
}
\value{
html upsetjs widget
}
\description{
upsetjs - factory for UpSet HTMLWidget
}
