% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsetjs.R
\name{upsetjsProxy}
\alias{upsetjsProxy}
\title{reactive helper to update an upsetjs inplace}
\usage{
upsetjsProxy(outputId, session)
}
\arguments{
\item{outputId}{the id of the upsetjs widget}

\item{session}{current shiny session}
}
\description{
reactive helper to update an upsetjs inplace
}
