% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{chartLabels}
\alias{chartLabels}
\title{specify chart labels}
\usage{
chartLabels(upsetjs, combinations.y.label = NULL, sets.x.label = NULL)
}
\arguments{
\item{upsetjs}{the upsetjs (proxy) instance}

\item{combinations.y.label}{the label for the combination chart}

\item{sets.x.label}{the label for the set chart}
}
\description{
specify chart labels
}
