% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{chartLayout}
\alias{chartLayout}
\title{specify the chart layout}
\usage{
chartLayout(
  upsetjs,
  height.ratios = NULL,
  width.ratios = NULL,
  padding = NULL,
  bar.padding = NULL,
  dot.padding = NULL
)
}
\arguments{
\item{upsetjs}{the upsetjs (proxy) instance}

\item{height.ratios}{a vector of length 2 for the ratios between the combination and set plot, e.g. c(0.6, 0.4)}

\item{width.ratios}{a vector of length 3 for the ratios between set, label, and combination plot, e.g. c(0.3,0.2,0.5)}

\item{padding}{padding around the plot}

\item{bar.padding}{padding ratio (default 0.1) for the bar charts}

\item{dot.padding}{padding facotr (default 0.7) for the dots}
}
\value{
upsetjs
}
\description{
specify the chart layout
}
