% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{chartStyleFlags}
\alias{chartStyleFlags}
\title{specify chart flags}
\usage{
chartStyleFlags(
  upsetjs,
  query.legend = NULL,
  export.buttons = NULL,
  class.name = NULL
)
}
\arguments{
\item{upsetjs}{the upsetjs (proxy) instance}

\item{query.legend}{show query legend if needed}

\item{export.buttons}{show export SVG and PNG buttons}

\item{class.name}{extra CSS class name to the root element}
}
\value{
upsetjs
}
\description{
specify chart flags
}
