% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{fromDataFrame}
\alias{fromDataFrame}
\title{extract the sets from a data frame (rows = elems, columns = sets, cell = contained)}
\usage{
fromDataFrame(upsetjs, df, order.by = "freq")
}
\arguments{
\item{upsetjs}{the upsetjs (proxy) instance}

\item{df}{the data.frame like structure}

\item{order.by}{order intersections by frequency (freq) or degree (deg)}
}
\description{
extract the sets from a data frame (rows = elems, columns = sets, cell = contained)
}
