% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{fromExpression}
\alias{fromExpression}
\title{generates the sets from a lists object that contained the cardinalties of both sets and combinations (+)}
\usage{
fromExpression(upsetjs, value, order.by = "freq")
}
\description{
generates the sets from a lists object that contained the cardinalties of both sets and combinations (+)
}
