% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{fromList}
\alias{fromList}
\title{generates the sets from a lists object}
\usage{
fromList(upsetjs, value, order.by = "cardinality")
}
\arguments{
\item{upsetjs}{the upsetjs (proxy) instance}

\item{value}{the list input value}

\item{order.by}{order intersections by cardinality or degree}
}
\description{
generates the sets from a lists object
}
