% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{generateIntersections}
\alias{generateIntersections}
\title{generate intersections}
\usage{
generateIntersections(
  upsetjs,
  min = NULL,
  max = NULL,
  empty = NULL,
  order.by = "freq",
  limit = NULL
)
}
\arguments{
\item{upsetjs}{the upsetjs (proxy) instance}

\item{min}{minimum number of sets in an intersection}

\item{max}{maximum number of sets in an intersection}

\item{empty}{whether to include empty intersections or not}

\item{order.by}{order intersections by frequency (freq) or degree (deg)}

\item{limit}{limit the number of intersections to the top N}
}
\description{
generate intersections
}
