% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{generateUnions}
\alias{generateUnions}
\title{configure the generatation of the unions}
\usage{
generateUnions(
  upsetjs,
  min = NULL,
  max = NULL,
  empty = NULL,
  order.by = "cardinality",
  limit = NULL
)
}
\arguments{
\item{upsetjs}{the upsetjs (proxy) instance}

\item{min}{minimum number of sets in an union}

\item{max}{maximum number of sets in an union}

\item{empty}{whether to include empty intersections or not}

\item{order.by}{order intersections by cardinality or degree}

\item{limit}{limit the number of intersections to the top N}
}
\value{
upsetjs
}
\description{
configure the generatation of the unions
}
