% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{generateUnions}
\alias{generateUnions}
\title{generate intersections}
\usage{
generateUnions(upsetjs, min = NULL, max = NULL, order.by = "freq")
}
\arguments{
\item{upsetjs}{the upsetjs (proxy) instance}
}
\description{
generate intersections
}
