% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{interactiveChart}
\alias{interactiveChart}
\title{interactive}
\usage{
interactiveChart(upsetjs, value = TRUE)
}
\arguments{
\item{upsetjs}{the upsetjs (proxy) instance}

\item{value}{whether to enable or disable}
}
\description{
interactive
}
