% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{queryLegend}
\alias{queryLegend}
\title{renders a legend for the queries}
\usage{
queryLegend(upsetjs, value = TRUE)
}
\arguments{
\item{upsetjs}{the upsetjs (proxy) instance}

\item{value}{whether to enable or disable}
}
\value{
upsetjs
}
\description{
renders a legend for the queries
}
