% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsetjs-package.R, R/upsetjs.R
\docType{package}
\name{upsetjs}
\alias{upsetjs}
\title{Upset.js}
\usage{
upsetjs(
  width = "100\%",
  height = NULL,
  elementId = NULL,
  sizingPolicy = upsetjsSizingPolicy()
)
}
\arguments{
\item{width}{width of the element}

\item{height}{height of the element}

\item{elementId}{unique element id}

\item{sizingPolicy}{htmlwidgets sizing policy object. Defaults to \code{\link{upsetjsSizingPolicy}()}}
}
\value{
html upsetjs widget
}
\description{
upsetjs a htmlwidget wrapper around UpSet.js (\url{https://upsetjs.netlify.com/})
}
\examples{
upsetjs() \%>\% fromList(list(a=c(1,2,3), b=c(2,3)))
}
